improve main_sm.rs

+++ state-machines/main/src/main_sm.rs
@@ -34,7 +34,7 @@ pub struct MainInstance<F: PrimeField64> {
 
 impl<F: PrimeField64> MainInstance<F> {
     const MAX_SEGMENT_ID: usize =
-        ((DEFAULT_MAX_STEPS + 1) as usize / MainTrace::<()>::NUM_ROWS) - 1;
+        (((DEFAULT_MAX_STEPS + 1) / MainTrace::<()>::NUM_ROWS as u64) - 1) as usize;
 
     /// Creates a new `MainInstance`.
     ///
